api/pyproject.toml
[project]
name = "wandlung"
version = "0.1.0"
description = "Wandlung REST api server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-storages>=1.14.4",
    "django>=5.1.4",
    "django-ninja>=1.3.0",
    "pillow>=11.0.0",
    "python-decouple>=3.8",
    "boto3>=1.35.90",
    "django-extensions>=3.2.3",
    "yt-dlp>=2024.12.23",
    "ffmpy>=0.5.0",
    "openai>=1.58.1",
]
